% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_date.R
\name{get_max_date}
\alias{get_max_date}
\title{get_max_date}
\usage{
get_max_date(filepath, components)
}
\arguments{
\item{filepath}{filepath}

\item{components}{components}
}
\value{

}
\description{
get_max_date
}
