% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journal_names}
\alias{journal_names}
\title{Conversion tables for journal names and abbreviations}
\format{
An object of class \code{data.frame} with 229148 rows and 2 columns.
}
\usage{
journal_names
}
\description{
Conversion tables for journal names and abbreviations
}
\keyword{datasets}
