% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cam_ambulance.R
\name{process_cam_ambulance}
\alias{process_cam_ambulance}
\title{process_cam_ambulance}
\usage{
process_cam_ambulance(sourcefile, filename)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}
}
\description{
Process a subset of the cases-and-management dataset
}
