% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cam_hospital.R
\name{process_cam_hospital}
\alias{process_cam_hospital}
\title{process_cam_hospital}
\usage{
process_cam_hospital(handle, input_path)
}
\arguments{
\item{handle}{list}

\item{input_path}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}
}
\description{
Process a subset of the cases-and-management dataset
}
