% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nrs_demographics.R
\name{process_nrs_demographics}
\alias{process_nrs_demographics}
\title{process_nrs_demographics}
\usage{
process_nrs_demographics(
  sourcefile,
  h5filename,
  grp.names,
  full.names,
  age.classes,
  conversionh5filepath,
  conversionh5version_number,
  genderbreakdown
)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{h5filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}

\item{grp.names}{a \code{string} specifying the shortened names of the 
administrative geographies - which should match those in the conversion table
 - and the sizes of the grid squares used in the conversion table in the 
 format gridxkm}

\item{full.names}{a \code{string} specifying the full names of the 
administrative geographies to which the data is converted to.}

\item{age.classes}{a \code{string} specifying the lower bounds of the age 
classes to which the data should be assigned.}

\item{conversionh5filepath}{a \code{string} specifying the local path
associated with the conversion table}

\item{conversionh5version_number}{a \code{string} specifying the version 
number (which is also the filename) of the conversion table}

\item{genderbreakdown}{a \code{string} specifying the names which should 
be assigned to gender categories in the output file and the gender files 
from the input which should be used in each categotry}
}
\description{
process_nrs_demographics
}
