% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pollution_data.R
\name{process_pollution_data}
\alias{process_pollution_data}
\title{process_pollution_data}
\usage{
process_pollution_data(
  source_filename,
  years,
  conversion_table,
  storage_path,
  h5filename
)
}
\arguments{
\item{source_filename}{a \code{string} specifying the local path associated 
with the folders containing the pollution source data (the input of this function)
which is separated by year}

\item{years}{a \code{string} specifying the years of pollution data which are 
to be collated}

\item{conversion_table}{a \code{string} specifying the local path and filename
associated with a lookup file of UK national grid codes and UKAIR pollution codes}

\item{storage_path}{a \code{string }specifying the local path associated 
with the processed data (the output of this function)}

\item{h5filename}{a \code{string} specifying the filename
associated with the processed data (the output of this function)}
}
\description{
process_pollution_data
}
