% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_lookup.R
\name{process_scotgov_lookup}
\alias{process_scotgov_lookup}
\title{process_scotgov_lookup}
\usage{
process_scotgov_lookup(sourcefile, h5filename, path, grid_names)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{h5filename}{a \code{string} specifying the filename
associated with the processed data (the output of this function)}

\item{path}{a \code{string} specifying the local path associated 
with the processed data}

\item{grid_names}{a \code{string} specifying the sizes of the grid squares
used in the conversion table in the format gridxkm}
}
\description{
process_scotgov_lookup
}
