% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_management.R
\name{process_scotgov_management}
\alias{process_scotgov_management}
\title{process_scotgov_management}
\usage{
process_scotgov_management(sourcefile, filename)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}
}
\description{
process_scotgov_management
}
