% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ukgov_eng_lookup.R
\name{process_ukgov_eng_lookup}
\alias{process_ukgov_eng_lookup}
\title{process_ukgov_eng_lookup}
\usage{
process_ukgov_eng_lookup(sourcefile, h5filename, output_area_sf, grid_names)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{h5filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}

\item{output_area_sf}{a \code{string} specifying the local path and filename
associated with the UK government output area shapefile}

\item{grid_names}{a \code{string} specifying the sizes of the grid squares 
used in the conversion table in the format gridxkm}
}
\description{
process_ukgov_eng_lookup
}
