% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journal_names}
\alias{journal_names}
\title{Conversion table for journal names and abbreviations}
\format{
A data frame with 2 columns
\describe{
  \item{journal}{full journal name}
  \item{abbreviation}{journal name abbreviation}
}
}
\usage{
journal_names
}
\description{
A dataset containing a conversion table for journal names and abbreviations
}
\keyword{datasets}
