% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cam_carehomes.R
\name{process_cam_carehomes}
\alias{process_cam_carehomes}
\title{process_cam_carehomes}
\usage{
process_cam_carehomes(handle, input_path)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{input_path}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}
}
\description{
Process a subset of the cases-and-management dataset
}
