% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ons_demographics.R
\name{process_ons_demographics}
\alias{process_ons_demographics}
\title{process_ons_demographics}
\usage{
process_ons_demographics(sourcefile, h5filename, h5path, conversionfile)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{h5filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}

\item{h5path}{a \code{string} specifying the local path
associated with the processed data (the output of this function)}

\item{conversionfile}{a \code{data.frame} containing a spatial conversion table}
}
\description{
process_ons_demographics
}
