% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pollution_lookup.R
\name{process_pollution_lookup}
\alias{process_pollution_lookup}
\title{process_pollution_lookup}
\usage{
process_pollution_lookup(sourcefile, filename, output_area_sf)
}
\arguments{
\item{sourcefile}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}

\item{filename}{a \code{string} specifying the local path and filename
associated with the processed data (the output of this function)}

\item{output_area_sf}{a \code{string} specifying the local path and filename
associated with a shapefile of England/Wales output areas}
}
\description{
process_pollution_lookup
}
