% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_deaths.R
\name{process_scotgov_deaths}
\alias{process_scotgov_deaths}
\title{process_scotgov_deaths}
\usage{
process_scotgov_deaths(handle, input_path)
}
\arguments{
\item{handle}{list}

\item{input_path}{a \code{string} specifying the local path and filename
associated with the source data (the input of this function)}
}
\description{
process_scotgov_deaths
}
