% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_lookup.R
\name{process_scotgov_lookup}
\alias{process_scotgov_lookup}
\title{process_ukgov_eng_lookup}
\usage{
process_scotgov_lookup(sourcefile, h5filename, grid_names, datazone_sf)
}
\description{
process_ukgov_eng_lookup
}
